refactor, move isLocalHost from GPAT to UrlUtil GP-5943
