+++ src/org/genepattern/server/genepattern/GenePatternAnalysisTask.java
@@ -125,6 +125,7 @@ import org.genepattern.server.database.HibernateSessionManager;
 import org.genepattern.server.database.HibernateUtil;
 import org.genepattern.server.dm.GpFileObjFactory;
 import org.genepattern.server.dm.GpFilePath;
+import org.genepattern.server.dm.UrlUtil;
 import org.genepattern.server.domain.AnalysisJob;
 import org.genepattern.server.domain.AnalysisJobDAO;
 import org.genepattern.server.domain.JobStatus;
@@ -295,96 +296,6 @@ public class GenePatternAnalysisTask {
             }
         }
         return inputFileMode;
-    }
-
-    /**
-     * utility call to get the IP address to the url host. 
-     * @param url, presumably a link to a data file
-     * @return an InetAddress, or null if errors occur
-     */
-    protected static InetAddress sys_requestedAddress(final URL url) {
-        try {
-            return InetAddress.getByName(url.getHost());
-        }
-        catch (UnknownHostException e) {
-            if (log.isDebugEnabled()) {
-                log.debug(e);
-            }
-        }
-        catch (Throwable t) {
-            log.error(t);
-        }
-        return null;
-    }
-
-    protected static boolean isLocalIpAddress(final InetAddress addr) {
-        if (addr==null) {
-            return false;
-        }
-
-        // Check if the address is a valid special local or loop back
-        if (addr.isAnyLocalAddress() || addr.isLoopbackAddress()) {
-            return true;
-        }
-
-        // Check if the address is defined on any interface
-        NetworkInterface ni=null;
-        try {
-            ni=NetworkInterface.getByInetAddress(addr);
-            return ni != null;
-        } 
-        catch (final SocketException e) {
-            if (log.isDebugEnabled()) {
-                log.debug("error in getByInetAddress, addr=", e);
-            }
-        }
-        catch (Throwable t) {
-            log.error("unexpected error in getByInetAddress, addr="+addr, t);
-        }
-        return false;
-    }
-    
-    protected static boolean isLocalIpAddress(final InetAddress addr, final NetworkInterface ni) {
-        if (ni != null) {
-            return true;
-        }
-        return false;
-    }
-
-    /**
-     * Tests whether the specified URL refers to the local host.
-     * 
-     * @param url, The URL to check whether it refers to the local host.
-     * @return <tt>true</tt> if the specified URL refers to the local host.
-     */
-    protected static boolean isLocalHost(final GpConfig gpConfig, final URL url) {
-        if (url==null) {
-            return false;
-        }
-        final String requestedHost=url.getHost();
-        if (requestedHost==null) {
-            return false;
-        }
-        // short-circuit test for 'localhost' and '127.0.0.1' to avoid invoking InetAddress methods
-        if (requestedHost.equals("localhost")) {
-            return true;
-        }
-        if (requestedHost.equals("127.0.0.1")) {
-            return true;
-        }
-        // check GpConfig.genePatternURL (from genepattern.properties)
-        final URL gpUrl = gpConfig.getGenePatternURL();
-        if (gpUrl != null && requestedHost.equals(gpUrl.getHost())) {
-            return true;
-        }
-        
-        final InetAddress requestedAddress=sys_requestedAddress(url);
-        if (requestedAddress != null) {
-            if (isLocalIpAddress(requestedAddress)) {
-                return true;
-            }
-        }
-        return false;
     }
 
     /**
@@ -1219,7 +1130,7 @@ public class GenePatternAnalysisTask {
                             }
                             else {
                                 final URL url = uri.toURL();
-                                final boolean isLocalHost=isLocalHost(gpConfig, url);
+                                final boolean isLocalHost=UrlUtil.isLocalHost(gpConfig, url);
                                 if (log.isDebugEnabled()) {
                                     log.debug("isLocalHost("+url+")="+isLocalHost);
                                 }
